+++ frontend/utils/src/index.ts
@@ -19,4 +19,4 @@ import "./polyfills"
 export * from "./browser"
 export * from "./types"
 export * from "./uri"
-export * from "./iframe/iframeSizing"
+export * from "./iframe/IframeSizer"
